requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
sentence-transformers==2.2.2
chromadb==0.4.18
pinecone-client==2.2.4
python-dotenv==1.0.0
PyPDF2==3.0.1
pymupdf==1.23.8
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
prometheus-client==0.19.0
structlog==23.2.0
tenacity==8.2.3
tiktoken==0.5.2
numpy==1.24.4
pandas==2.1.4
aiofiles==23.2.1
